+++ source/raw/bin/x64/plugins/cyber_engine_tweaks/mods/Character Preset Manager (CET)/modules/state.lua
@@ -155,6 +155,7 @@ state = {
     skipped = 0,
     requestSequence = 0,
     requested = false,
+    dangerousLogged = {},
   },
   ui = {
     openSections = {